requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0  # .env dosyası desteği
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0